check working sensor communication in init()
Previously, init() behaved differently depending on
whether SHTSensor was instantiated with an fixed sensor
type or not (auto-detect). In the auto-detect case, it would
trigger a 'readSample()' and thus ensure that the
sensor was communicating, while the fixed initialization
no such check was performed.

This change addressed this by calling readSample() in both cases,
and thus guaranteeing that if init() returns true, the sensor driver
was successfully created, and communication with the sensor on the bus
has been established